pyproject.toml
[tool.poetry]
name = "bank-api"
version = "1.0.0"
description = "API Bancária Assíncrona com FastAPI"
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = ">=0.111.0"
uvicorn = {extras = ["standard"], version = ">=0.29.0"}
PyJWT = ">=2.7.0"
passlib = {extras = ["bcrypt"], version = ">=1.7.4"}
bcrypt = "4.0.1"
sqlalchemy = ">=2.0.0"
aiosqlite = ">=0.20.0"
pydantic = ">=2.0.0"
pydantic-settings = ">=2.0.0"
alembic = ">=1.13.0"
python-multipart = ">=0.0.9"
greenlet = "^3.3.2"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"
anyio = ">=4.0.0"
httpx = ">=0.27.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
